Big Changes
I realized there's this wonderful thing call "inherticance" and instead of clogging up the player class with all types of junk, I can inherit the properites of a player into a gatheringCalc and a craftingCalc (not yet implimented) and augment their methods from there.

In addition, bonuses work and I added in Mastery prediction, so it will level up at the next tick interval and apply the bonuses that come with it. Next step...equiptment...somehow... it should be much easier with the rewrite in place.

I will eventually incorperate everything into TalibriCalc, but for now everything is in TalibriCalcRewrite